User story 3 complete
