--- spec/meal_spec.rb
@@ -1,9 +1,0 @@
-require 'meal'
-require 'menu'
-
-describe Meal do
-  it 'stores selected dishes' do
-    subject.add("Fish")
-    expect(subject.meal).to eq(["Fish"])
-  end
-end
